+++ fixtures/transform/template-token.ts
@@ -1,0 +1,4 @@
+const x = `1${2}3`;
+
+// @dts-jest
+console.log(x);