+++ src/__tests__/transform.ts
@@ -12,6 +12,10 @@ it('should transform correctly', () => {
   expect(transform_fixture('../../fixtures/transform/general.ts')).toMatchSnapshot();
 });
 
+it('should transform correctly with template token', () => {
+  expect(transform_fixture('../../fixtures/transform/template-token.ts')).toMatchSnapshot();
+});
+
 it('should throw error with unexpected flag', () => {
   expect(() => transform_fixture('../../fixtures/transform/unexpected.ts')).toThrowError();
 });